+++ lib/impl/atomizer/any.ex
@@ -1,3 +1,3 @@
 defimpl Digger.Atomizer, for: Any do
-  def atomize(any, _atomize), do: any
+  def atomize(any, _opts), do: any
 end