+++ lib/impl/atomizer/list.ex
@@ -1,11 +1,9 @@
 defimpl Digger.Atomizer, for: List do
   alias Digger.Atomizer
 
-  def atomize([], _atomize), do: []
+  def atomize([], _opts), do: []
 
-  def atomize([h | t], :atomize) do
-    [Atomizer.atomize(h, :atomize)] ++ Atomizer.atomize(t, :atomize)
+  def atomize([h | t], opts) do
+    [Atomizer.atomize(h, opts)] ++ Atomizer.atomize(t, opts)
   end
-
-  def atomize(list, _atomize), do: list
 end