+++ lib/impl/atomizer/map.ex
@@ -1,8 +1,17 @@
 defimpl Digger.Atomizer, for: Map do
   alias Digger.Atomizer
 
-  def atomize(map, _atomize) do
+  def atomize(map, type: _type, atomize_key: atomize_key, atomize_value: atomize_value) do
     map
-    |> Enum.reduce(%{}, fn({key, value}, acc) -> Map.merge(acc, %{Atomizer.atomize(key, :atomize) => Atomizer.atomize(value, :no_atomize)}) end)
+    |> Enum.reduce(%{}, fn {key, value}, acc ->
+      Map.merge(acc, %{
+        Atomizer.atomize(key, type: :key, atomize_key: atomize_key, atomize_value: atomize_value) =>
+          Atomizer.atomize(value,
+            type: :value,
+            atomize_key: atomize_key,
+            atomize_value: atomize_value
+          )
+      })
+    end)
   end
 end