+++ lib/protocols/atomizer_protocol.ex
@@ -11,6 +11,6 @@ defprotocol Digger.Atomizer do
   'Atomize' a valid Types.data_type according to the protocol implementation
   """
 
-  @spec atomize(Types.data_type, Types.string) :: Types.valid_return_type
-  def atomize(data_type, atomize \\ :atomize)
+  @spec atomize(Types.data_type(), keyword()) :: Types.valid_return_type()
+  def atomize(data_type, opts \\ [type: :key, atomize_key: true, atomize_value: false])
 end