+++ CHANGELOG.md
@@ -10,6 +10,7 @@ For more information about changelogs, check
 
 -   [breaking change] - Stringifier module API updated to allow maps nested within lists to have their keys (and values) "stringified".
 -   [breaking change] - CamelCaser module API updated to allow maps nested within lists to have their keys (and values) "camelized".
+-   [breaking change] - Atomizer module API updated to allow maps nested within lists to have their keys (and values) "atomized".
 
 ## 2.0.1 - 12/12/18 Wed